// src/lib/interfaces/menu-event-definition.ts
import { MenuItemEventType } from '../enums';
import { MenuIdentifier } from './menu-identifier';

/**
 * MenuEventDefinition object Interface
 */
export interface MenuEventDefinition {
  type: MenuItemEventType;
  source: MenuIdentifier;
  target?: MenuIdentifier;
  data?: Record<string, string | number | boolean>;
}